refactor!: separate pulumi and configuration construct from the actual infrastructure
