+++ capeinfra/pipeline/batch.py
@@ -3,8 +3,9 @@
 import pulumi_aws as aws
 from pulumi import Input, ResourceOptions
 
+from capepulumi import CapeComponentResource
+
 from ..iam import get_inline_role, get_instance_profile
-from ..resources.pulumi import CapeComponentResource
 
 
 class BatchCompute(CapeComponentResource):
