+++ tests/conftest.py
@@ -35,12 +35,11 @@ def config_init():
 
 @pytest.fixture(scope="module")
 def mock_meta(mock_namespace):
+    import capeinfra
     from capeinfra.meta.capemeta import CapeMeta
-    from capeinfra.resources.pulumi import CapeComponentResource
 
-    meta = CapeMeta(f"{mock_namespace}-meta")
-    CapeComponentResource.meta = meta
-    return meta
+    capeinfra.meta = CapeMeta(f"{mock_namespace}-meta")
+    return capeinfra.meta
 
 
 @pytest.fixture(scope="module")
